requirements.txt
beautifulsoup4
dnspython3 
requests
python-geoip
